refactor: make code mode readable and abstract
